requirements.txt
Flask>=3.1.0
python-dotenv>=1.1.0
flask-cors>=5.0.0
flask-limiter>=3.9.0
flask-migrate>=4.0.0
flask-talisman>=1.1.0
flask-wtf>=1.2.0
flask-jwt-extended>=4.7.0
flask-sqlalchemy>=3.1.0
python-json-logger>=2.0.7
gunicorn>=23.0.0
httpx>=0.28.1
asgiref>=3.8.1
docker>=7.1.0
requests>=2.31.0
